Refactor: Extract entry migrations module
Move config-entry migration logic into a dedicated module while
preserving the package-level Home Assistant entry point. Split the
migration flow into per-version helpers so each transition remains
small and reviewable without changing migration semantics.

Closes #572
